1) Rebuild with maven 2) small refactoring
