+++ src/main/java/com/vandd/solutions/maze/algorithms/pathfind/FindingExit.java
@@ -1,8 +1,8 @@
-package Algorithms;
+package com.vandd.solutions.maze.algorithms.pathfind;
 
-import GridModel.Grid;
-import GridModel.Tile;
-import GridModel.Painter;
+import com.vandd.solutions.maze.GridModel.Grid;
+import com.vandd.solutions.maze.GridModel.Tile;
+import com.vandd.solutions.maze.GridModel.Painter;
 import java.util.List;
 public abstract class FindingExit {
     protected Painter painter;
