+++ src/main/resources/com/vandd/solutions/maze/autumn.jpg
