+++ src/main/resources/com/vandd/solutions/maze/icon.png
