+++ src/main/resources/com/vandd/solutions/maze/spring.jpg
