+++ src/main/resources/com/vandd/solutions/maze/summer.jpg
