+++ src/main/resources/com/vandd/solutions/maze/winter.jpg
